fix(utils): scrub macOS malloc env from child spawns (#693)
